requirements.txt
absl-py==1.4.0
beautifulsoup4==4.12.2
black==23.3.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
filelock==3.12.0
fonttools==4.39.3
gdown==4.7.1
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
protobuf==4.22.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
rdkit==2023.3.1
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
ruff==0.0.265
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==2.0.0
torch-geometric==2.3.1
torchvision==0.15.1
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.3
zipp==3.15.0
